Allow certain sysctl parameters to be overridden
Tuning applications on top of Taupage sometimes requires tuning certain (kernel)
parameters. This allows Taupage configurations to have a sysctl section which
will be used instead of the defaults.
As some kernel parameters may not be allowed to be tuned, only parameters
on a whitelist are allowed to be specified.

This addresses issues:
* https://github.com/zalando-stups/taupage/issues/175
* https://github.com/zalando/spilo/issues/42