Start docker container if it already exists. Fix #199
This is a workaround for an issue with startup sequence when we
have external volumes to mount.  In case of EC2 instance restart
the Docker daemon will start prior to taupage-init and will
immediately try to start the container it was running prior to
the restart.  This will fail because the volumes are not there yet,
they are mounted as part of taupage-init script, in 10-prepare-disks.py.

Starting the container explicitly "fixes" the problem because at that
point all required initialization steps are completed.