Resize mounts by default
Volumes that should not be erased_on_boot may have been created
from different snapshots. One of the reasons to do this is to be able
to scale data stores (PostgreSQL, Cassandra) once they fill up.
What happens in this case is that the volume is resized, however the
filesystem still has the previous size.
Resizing the filesystem to the volume seems like a win, we can do
it unconditionally as resize2fs will figure out by itself whether
a resize is needed.

Addresses issue #252