Add support for attaching network interfaces based of a tag
This feature enables you to associate a Static IP to an instance.

Given a network interface exists within the same AZ as your instance,
and the network interface has a name tag, the value of the name tag
can be specified in the taupage config and taupage will then attach
and configure it on boot up.

```
network_interfaces:
    - Name_tag_value
```

This is helpful for applications like ZooKeeper that do not play
nicely behind a load balancer: https://issues.apache.org/jira/browse/ZOOKEEPER-1846

Signed-off-by: Ian Duffy <ian.duffy@zalando.ie>