Implement trusted image validation
Add a simple check if the image is trusted in the Docker registry before
running it. This is done by parsing the image tag and checking for the
X-Trusted: true header in the image manifest response. Note that this
only applies to the "main" container that's started by Taupage,
enforcing it for all pulls would break a lot of things like CI
containers.
An additional benefit is that this change would also make it impossible
to run images from registries other than Pier One since they don't send
the header in the first place. This is already a compliance requirement,
but there's a bit more enforcement now.