Increase number of attempts to find EBS volume
If we are running in ASG and instance got unhealthy, ASG will first
remove it from the group, spawn a replacement. The old instance will be
terminated in only in about 10 minutes.
As a result, the new instance is failing to attach EBS volumes because
they are still attached to the old instance.
In order to fix this problem I increase number of attempts to find the
volumes to 10 and increase the timeout between attempts to 2 minutes.

Signed-off-by: Alexander Kukushkin <alexander.kukushkin@zalando.de>