Implement trusted image validation (#505)
* Implement trusted image validation (#502)

* Implement trusted image validation

Add a simple check if the image is trusted in the Docker registry before
running it. This is done by parsing the image tag and checking for the
X-Trusted: true header in the image manifest response. Note that this
only applies to the "main" container that's started by Taupage,
enforcing it for all pulls would break a lot of things like CI
containers.
An additional benefit is that this change would also make it impossible
to run images from registries other than Pier One since they don't send
the header in the first place. This is already a compliance requirement,
but there's a bit more enforcement now.

* enforce images to be trusted before running them (#504)

* enforce images to be trusted before running them

* fix: use correct comparison with None

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* refrain from using pierone api client to get trusted image info

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* fix: use pierone's iid_auth() instead of making the call ourselves

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* Support the new trust status headers

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* Fix flake8 errors

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>

* Ignore X-Trusted if X-Production-Ready-Taupage is set to false

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>