Allow IPv4 address substitutions in EBS volume names
Fixes #488

Limit scope of substitutions to EBS volume names only.  This is not needed in
any other places of configuration (otherwise there would be open issues for
those as well).

Only allow `local-ipv4` and `public-ipv4` vars from instance metadata: for the
volume name substitutions these are the most useful ones.

Signed-off-by: Oleksandr Shulgin <oleksandr.shulgin@zalando.de>